requirements.txt
# Core dependencies
pymongo==4.5.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
schedule==1.2.0
Flask==2.3.3
Flask-CORS==4.0.0

# Testing dependencies
pytest==7.4.3
pytest-mock==3.12.0

# Development dependencies
black==23.10.1
flake8==6.1.0
isort==5.12.0
